Backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.3.2
cloudinary==1.43.0
cryptography==43.0.1
Django==5.1.1
django-admin-logs==1.3.0
django-admin-rangefilter==0.13.2
django-appconf==1.0.6
django-cloudinary-storage==0.3.0
django-email==0.1.10
django-jazzmin==3.0.0
django-mail-admin==0.3.2
django-select2==8.2.1
django-smtp-ssl==1.0
djangorestframework==3.16.0
et-xmlfile==1.1.0
idna==3.10
jsonfield==3.1.0
numpy==2.1.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pdfminer.six==20240706
pillow==10.4.0
psycopg2==2.9.10
pycparser==2.22
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.3.0
